fix admin`s delete

+++ controllers/admin.controllers.js
@@ -1,11 +1,12 @@
 const { instanceTransaction } = require('../dataBase').getInstance();
 const { statusCodes, statusMessages, emailActionsEnum } = require('../constants');
+const { ErrorHandler } = require('../helpers');
 
 // const { userService, mailService } = require('../services');
 const userService = require('../services/user.services')
 const mailService = require('../services/mail.services')
 
-const getAllUsers = () => {};
+const getBlockedUsers = () => {};
 
 const changeUserStatus = () => {};
 
@@ -15,6 +16,10 @@ const deleteUser = async (req, res, next) => {
     const { params: { userID }, query: { prefLang = 'en' } } = req;
 
     const { name, email } = await userService.findUserById(userID) || {};
+
+    if (!email) {
+      throw new ErrorHandler(statusMessages.USERS_NOT_FOUND[prefLang], statusCodes.BAD_REQUEST)
+    }
 
     await userService.deleteUser(userID, transaction);
 
@@ -29,7 +34,7 @@ const deleteUser = async (req, res, next) => {
   }
 }
 module.exports = {
-  getAllUsers,
+  getBlockedUsers,
   changeUserStatus,
   deleteUser
 }
